feat: add SQL indexes for badges, plan activities, and user activity queries; update application configuration and Dockerfile

+++ progress-service/src/main/resources/db/migration/V2__add_missing_stats_columns.sql
@@ -3,4 +3,4 @@ ALTER TABLE user_stats
     ADD COLUMN IF NOT EXISTS total_chat_sessions INT NOT NULL DEFAULT 0,
     ADD COLUMN IF NOT EXISTS last_active_date BIGINT,
     ADD COLUMN IF NOT EXISTS streak_start_date BIGINT,
-    ADD COLUMN IF NOT EXISTS average_mood_score DECIMAL(4,2);
+    ADD COLUMN IF NOT EXISTS average_mood_score DOUBLE PRECISION;
